# docs/openapi.yaml
openapi: 3.0.3
info:
  title: E-commerce Backend API
  version: 1.0.0
  description: |
    OpenAPI specification generated from the source code in this repository.

    Notes:
    - Base path has no `/api` prefix.
    - This spec reflects the current code, including route naming inconsistencies.
    - Many admin-style routes are not protected in the backend.
    - `GET /getsinglesubcat` is documented as implemented, even though the route is misconfigured.
servers:
  - url: http://localhost:5001
    description: Local development server
...
paths:
...
  /getsignlecategory/{id}:
    get:
      tags: [Categories]
      summary: Get a category by id
      operationId: getSingleCategory
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Category result array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /updatecategory/{id}:
    patch:
      tags: [Categories]
      summary: Update a category name
      operationId: updateCategory
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category_name]
              properties:
                category_name:
                  type: string
              example:
                category_name: Consumer Electronics
      responses:
        '200':
          description: Update count array
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateCountArray'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /deletecategory/{id}:
    delete:
      tags: [Categories]
      summary: Delete a category
      operationId: deleteCategory
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Category deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericMessage'
              example:
                message: Category Deleted Succesfully
        '400':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericMessage'
              example:
                message: No such Category found.
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
      example: 1
...
  responses:
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericMessage'
          example:
            message: Internal server error
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
        category_name:
          type: string
        uploadby:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        id: 1
        category_name: Electronics
        uploadby: admin@example.com
        createdAt: '2024-01-01T00:00:00.000Z'
        updatedAt: '2024-01-01T00:00:00.000Z'
...
    GenericMessage:
      type: object
      properties:
        message:
          type: string
      example:
        message: Operation completed
...
    UpdateCountArray:
      type: array
      items:
        type: integer
      example: [1]
...